DBapplicationDownloader_v110
## Auto downloader
* Based on flag that written by the ap as a False by default and can be changeable by an external .bat file for update when this app is running - and open the latest modified application in the application folder.

## Logger
* logger.debug("Message"): Log a debug message (level 10).
* logger.info("Message"): Log an informational message (level 20).
* logger.error("Message"): Log an error message (level 40).
* In case of unhandled exceptions, the logger will automatically log the details.
